feat: Move BadgeUtils to shared, add API to get badge info
